fix: 포인트 사용시 파라미터 타입 수정
파라미터(usePoint)가 int로 되어 있어 long으로 수정함.

+++ src/main/java/io/hhplus/tdd/point/UserPointFacade.java
@@ -29,7 +29,7 @@ public class UserPointFacade {
         }
     }
 
-    public UserPoint usePoint(long id, int usePoint) {
+    public UserPoint usePoint(long id, long usePoint) {
         UserPoint originUserPoint = userPointService.getPoint(id);
         try {
             UserPoint chargedUserPoint = userPointService.usePoint(id, usePoint);
